Revert "refactor: DBPT-1520-Refactoring conduit command and tests (part 1) (#623)"
This reverts commit 394077629ac1c58e99356c8b3ebe4023846f1a41.
