+++ README.md
@@ -69,9 +69,7 @@ We use [Codecov](https://app.codecov.io/github/uktrade/platform-tools) to monito
 
 You may want to test any CLI changes locally.
 
-##### Option 1 - Build and install `platform-helper` from your local source code
-
-Run `poetry build` to build your package resulting in a package file (e.g. `dbt_platform_tools-0.1.40.tar.gz`) in a `dist` folder. You may need to bump up the package version before doing so. To bump the version go to `pyproject.toml/version`
+Run `poetry build` to build your package resulting in a package file (e.g. `dbt_platform_tools-0.1.40.tar.gz`) in a `dist` folder. You may need to bump up the package version before doing so.
 
 Copy the package file(s) to the directory where you would like to test your changes, and make sure you are in a virtual environment. Run `platform-helper --version` to check the installed package version (e.g. `0.1.39`).
 
@@ -82,19 +80,6 @@ Run `pip install <file>` and confirm the installation has worked by running `pla
 
 > [!IMPORTANT]
 > When testing is complete, do not forget to revert the `dbt-platform-helper` installation back to what it was; e.g. `pip install dbt-platform-helper==0.1.39`.
-
-##### Option 2 - Run the python files directly.
-
-This assumes that the virtual python environment where you are running them from already has the dependencies installed and the directory is at the same level as your platform-tools directory.
-
-Example usage:
-
-```
-# From <application>-deploy
-
-../platform-tools/platform_helper.py <command> <options>
-```
-
 
 #### End to end testing
