+++ tests/platform_helper/fixtures/test_cloudformation_template.yml
@@ -1,6 +1,3 @@
-Parameters:
-  ExistingParameter:
-    Type: String
 Resources:
   DefaultTaskRole:
     Type: 'AWS::IAM::Role'
